Addon Manager: Enable testing utilities from IDE

+++ src/Mod/AddonManager/addonmanager_utilities.py
@@ -35,18 +35,23 @@ from typing import Optional, Any
 
 from urllib.parse import urlparse
 
-from PySide import QtCore, QtWidgets
+try:
+    from PySide import QtCore, QtWidgets
+except ImportError:
+    QtCore = None
+    QtWidgets = None
 
-import FreeCAD
+import addonmanager_freecad_interface as fci
 
-if FreeCAD.GuiUp:
+if fci.gui_up:
     import FreeCADGui
 
-    # If the GUI is up, we can use the NetworkManager to handle our downloads. If there is no event
-    # loop running this is not possible, so fall back to requests (if available), or the native
-    # Python urllib.request (if requests is not available).
+    # If the GUI is up, we can use the NetworkManager to handle our downloads. If
+    # there is no event loop running this is not possible, so fall back to requests (
+    # if available), or the native Python urllib.request (if requests is not available).
     import NetworkManager  # Requires an event loop, so is only available with the GUI
 else:
+    FreeCADGui = None
     try:
         import requests
     except ImportError:
@@ -60,7 +65,7 @@ else:
 #  @{
 
 
-translate = FreeCAD.Qt.translate
+translate = fci.translate
 
 
 class ProcessInterrupted(RuntimeError):
@@ -68,8 +73,8 @@ class ProcessInterrupted(RuntimeError):
 
 
 def symlink(source, link_name):
-    """Creates a symlink of a file, if possible. Note that it fails on most modern Windows
-    installations"""
+    """Creates a symlink of a file, if possible. Note that it fails on most modern
+    Windows installations"""
 
     if os.path.exists(link_name) or os.path.lexists(link_name):
         pass
@@ -78,9 +83,9 @@ def symlink(source, link_name):
         if callable(os_symlink):
             os_symlink(source, link_name)
         else:
-            # NOTE: This does not work on most normal Windows 10 and later installations, unless
-            # developer mode is turned on. Make sure to catch any exception thrown and have a
-            # fallback plan.
+            # NOTE: This does not work on most normal Windows 10 and later
+            # installations, unless developer mode is turned on. Make sure to catch
+            # any exception thrown and have a fallback plan.
             csl = ctypes.windll.kernel32.CreateSymbolicLinkW
             csl.argtypes = (ctypes.c_wchar_p, ctypes.c_wchar_p, ctypes.c_uint32)
             csl.restype = ctypes.c_ubyte
@@ -115,10 +120,6 @@ def update_macro_details(old_macro, new_macro):
     one the wiki.
     """
 
-    if old_macro.on_git and new_macro.on_git:
-        FreeCAD.Console.PrintLog(
-            f'The macro "{old_macro.name}" is present twice in github, please report'
-        )
     # We don't report macros present twice on the wiki because a link to a
     # macro is considered as a macro. For example, 'Perpendicular To Wire'
     # appears twice, as of 2018-05-05).
@@ -129,10 +130,11 @@ def update_macro_details(old_macro, new_macro):
 
 
 def remove_directory_if_empty(dir_to_remove):
-    """Remove the directory if it is empty, with one exception: the directory returned by
-    FreeCAD.getUserMacroDir(True) will not be removed even if it is empty."""
+    """Remove the directory if it is empty, with one exception: the directory
+    returned by FreeCAD.getUserMacroDir(True) will not be removed even if it is
+    empty."""
 
-    if dir_to_remove == FreeCAD.getUserMacroDir(True):
+    if dir_to_remove == fci.DataPaths().macro_dir:
         return
     if not os.listdir(dir_to_remove):
         os.rmdir(dir_to_remove)
@@ -140,6 +142,9 @@ def remove_directory_if_empty(dir_to_remove):
 
 def restart_freecad():
     """Shuts down and restarts FreeCAD"""
+
+    if not FreeCADGui or not QtWidgets or not QtCore:
+        return
 
     args = QtWidgets.QApplication.arguments()[1:]
     if FreeCADGui.getMainWindow().close():
@@ -156,17 +161,12 @@ def get_zip_url(repo):
         return f"{repo.url}/archive/{repo.branch}.zip"
     if parsed_url.netloc in ["gitlab.com", "framagit.org", "salsa.debian.org"]:
         return f"{repo.url}/-/archive/{repo.branch}/{repo.name}-{repo.branch}.zip"
-    FreeCAD.Console.PrintLog(
-        "Debug: addonmanager_utilities.get_zip_url: Unknown git host fetching zip URL:"
-        + parsed_url.netloc
-        + "\n"
-    )
     return f"{repo.url}/-/archive/{repo.branch}/{repo.name}-{repo.branch}.zip"
 
 
 def recognized_git_location(repo) -> bool:
-    """Returns whether this repo is based at a known git repo location: works with github, gitlab,
-    framagit, and salsa.debian.org"""
+    """Returns whether this repo is based at a known git repo location: works with
+    github, gitlab, framagit, and salsa.debian.org"""
 
     parsed_url = urlparse(repo.url)
     return parsed_url.netloc in [
@@ -185,11 +185,6 @@ def construct_git_url(repo, filename):
         return f"{repo.url}/raw/{repo.branch}/{filename}"
     if parsed_url.netloc in ["gitlab.com", "framagit.org", "salsa.debian.org"]:
         return f"{repo.url}/-/raw/{repo.branch}/{filename}"
-    FreeCAD.Console.PrintLog(
-        "Debug: addonmanager_utilities.construct_git_url: Unknown git host:"
-        + parsed_url.netloc
-        + f" for file {filename}\n"
-    )
     # Assume it's some kind of local GitLab instance...
     return f"{repo.url}/-/raw/{repo.branch}/{filename}"
 
@@ -207,19 +202,14 @@ def get_metadata_url(url):
 
 
 def get_desc_regex(repo):
-    """Returns a regex string that extracts a WB description to be displayed in the description
-    panel of the Addon manager, if the README could not be found"""
+    """Returns a regex string that extracts a WB description to be displayed in the
+    description panel of the Addon manager, if the README could not be found"""
 
     parsed_url = urlparse(repo.url)
     if parsed_url.netloc == "github.com":
         return r'<meta property="og:description" content="(.*?)"'
     if parsed_url.netloc in ["gitlab.com", "salsa.debian.org", "framagit.org"]:
         return r'<meta.*?content="(.*?)".*?og:description.*?>'
-    FreeCAD.Console.PrintLog(
-        "Debug: addonmanager_utilities.get_desc_regex: Unknown git host:",
-        repo.url,
-        "\n",
-    )
     return r'<meta.*?content="(.*?)".*?og:description.*?>'
 
 
@@ -231,37 +221,39 @@ def get_readme_html_url(repo):
         return f"{repo.url}/blob/{repo.branch}/README.md"
     if parsed_url.netloc in ["gitlab.com", "salsa.debian.org", "framagit.org"]:
         return f"{repo.url}/-/blob/{repo.branch}/README.md"
-    FreeCAD.Console.PrintLog(
-        "Unrecognized git repo location '' -- guessing it is a GitLab instance..."
-    )
     return f"{repo.url}/-/blob/{repo.branch}/README.md"
 
 
 def is_darkmode() -> bool:
     """Heuristics to determine if we are in a darkmode stylesheet"""
+    if not FreeCADGui:
+        return False
     pl = FreeCADGui.getMainWindow().palette()
     return pl.color(pl.Background).lightness() < 128
 
 
 def warning_color_string() -> str:
-    """A shade of red, adapted to darkmode if possible. Targets a minimum 7:1 contrast ratio."""
+    """A shade of red, adapted to darkmode if possible. Targets a minimum 7:1
+    contrast ratio."""
     return "rgb(255,105,97)" if is_darkmode() else "rgb(215,0,21)"
 
 
 def bright_color_string() -> str:
-    """A shade of green, adapted to darkmode if possible. Targets a minimum 7:1 contrast ratio."""
+    """A shade of green, adapted to darkmode if possible. Targets a minimum 7:1
+    contrast ratio."""
     return "rgb(48,219,91)" if is_darkmode() else "rgb(36,138,61)"
 
 
 def attention_color_string() -> str:
-    """A shade of orange, adapted to darkmode if possible. Targets a minimum 7:1 contrast ratio."""
+    """A shade of orange, adapted to darkmode if possible. Targets a minimum 7:1
+    contrast ratio."""
     return "rgb(255,179,64)" if is_darkmode() else "rgb(255,149,0)"
 
 
 def get_assigned_string_literal(line: str) -> Optional[str]:
-    """Look for a line of the form my_var = "A string literal" and return the string literal.
-    If the assignment is of a floating point value, that value is converted to a string
-    and returned. If neither is true, returns None."""
+    """Look for a line of the form my_var = "A string literal" and return the string
+    literal. If the assignment is of a floating point value, that value is converted
+    to a string and returned. If neither is true, returns None."""
 
     string_search_regex = re.compile(r"\s*(['\"])(.*)\1")
     _, _, after_equals = line.partition("=")
@@ -274,8 +266,8 @@ def get_assigned_string_literal(line: str) -> Optional[str]:
 
 
 def get_macro_version_from_file(filename: str) -> str:
-    """Get the version of the macro from a local macro file. Supports strings, ints, and floats,
-    as well as a reference to __date__"""
+    """Get the version of the macro from a local macro file. Supports strings, ints,
+    and floats, as well as a reference to __date__"""
 
     date = ""
     with open(filename, errors="ignore", encoding="utf-8") as f:
@@ -291,19 +283,10 @@ def get_macro_version_from_file(filename: str) -> str:
                 if match:
                     return match
                 if "__date__" in line.lower():
-                    # Don't do any real syntax checking, just assume the line is something
-                    # like __version__ = __date__
+                    # Don't do any real syntax checking, just assume the line is
+                    # something like __version__ = __date__
                     if date:
                         return date
-                    # pylint: disable=line-too-long,consider-using-f-string
-                    FreeCAD.Console.PrintWarning(
-                        translate(
-                            "AddonsInstaller",
-                            "Macro {} specified '__version__ = __date__' prior to setting a value for __date__".format(
-                                filename
-                            ),
-                        )
-                    )
             elif line.lower().startswith("__date__"):
                 match = get_assigned_string_literal(line)
                 if match:
@@ -315,11 +298,11 @@ def update_macro_installation_details(repo) -> None:
     """Determine if a given macro is installed, either in its plain name,
     or prefixed with "Macro_" """
     if repo is None or not hasattr(repo, "macro") or repo.macro is None:
-        FreeCAD.Console.PrintLog("Requested macro details for non-macro object\n")
+        fci.Console.PrintLog("Requested macro details for non-macro object\n")
         return
-    test_file_one = os.path.join(FreeCAD.getUserMacroDir(True), repo.macro.filename)
+    test_file_one = os.path.join(fci.DataPaths().macro_dir, repo.macro.filename)
     test_file_two = os.path.join(
-        FreeCAD.getUserMacroDir(True), "Macro_" + repo.macro.filename
+        fci.DataPaths().macro_dir, "Macro_" + repo.macro.filename
     )
     if os.path.exists(test_file_one):
         repo.updated_timestamp = os.path.getmtime(test_file_one)
@@ -342,9 +325,6 @@ def is_float(element: Any) -> bool:
         return False
 
 
-#  @}
-
-
 def get_python_exe() -> str:
     """Find Python. In preference order
     A) The value of the PythonExecutableForPip user preference
@@ -352,9 +332,9 @@ def get_python_exe() -> str:
     C) The executable located in the same bin directory as FreeCAD and called "python"
     D) The result of a shutil search for your system's "python3" executable
     E) The result of a shutil search for your system's "python" executable"""
-    prefs = FreeCAD.ParamGet("User parameter:BaseApp/Preferences/Addons")
+    prefs = fci.ParamGet("User parameter:BaseApp/Preferences/Addons")
     python_exe = prefs.GetString("PythonExecutableForPip", "Not set")
-    fc_dir = FreeCAD.getHomePath()
+    fc_dir = fci.DataPaths().home_dir
     if not python_exe or python_exe == "Not set" or not os.path.exists(python_exe):
         python_exe = os.path.join(fc_dir, "bin", "python3")
         if "Windows" in platform.system():
@@ -383,26 +363,26 @@ def get_pip_target_directory():
     # Get the default location to install new pip packages
     major, minor, _ = platform.python_version_tuple()
     vendor_path = os.path.join(
-        FreeCAD.getUserAppDataDir(), "AdditionalPythonPackages", f"py{major}{minor}"
+        fci.DataPaths().mod_dir, "..", "AdditionalPythonPackages", f"py{major}{minor}"
     )
     return vendor_path
 
 
 def get_cache_file_name(file: str) -> str:
     """Get the full path to a cache file with a given name."""
-    cache_path = FreeCAD.getUserCachePath()
+    cache_path = fci.DataPaths().cache_dir
     am_path = os.path.join(cache_path, "AddonManager")
     os.makedirs(am_path, exist_ok=True)
     return os.path.join(am_path, file)
 
 
 def blocking_get(url: str, method=None) -> str:
-    """Wrapper around three possible ways of accessing data, depending on the current run mode and
-    Python installation. Blocks until complete, and returns the text results of the call if it
-    succeeded, or an empty string if it failed, or returned no data. The method argument is
-    provided mainly for testing purposes."""
+    """Wrapper around three possible ways of accessing data, depending on the current
+    run mode and Python installation. Blocks until complete, and returns the text
+    results of the call if it succeeded, or an empty string if it failed, or returned
+    no data. The method argument is provided mainly for testing purposes."""
     p = ""
-    if FreeCAD.GuiUp and method is None or method == "networkmanager":
+    if fci.gui_up and method is None or method == "networkmanager":
         NetworkManager.InitializeNetworkManager()
         p = NetworkManager.AM_NETWORK_MANAGER.blocking_get(url)
     elif requests and method is None or method == "requests":
@@ -441,26 +421,25 @@ def run_interruptable_subprocess(args) -> subprocess.CompletedProcess:
             stdout, stderr = p.communicate(timeout=0.1)
             return_code = p.returncode
         except subprocess.TimeoutExpired:
-            if QtCore.QThread.currentThread().isInterruptionRequested():
+            if QtCore and QtCore.QThread.currentThread().isInterruptionRequested():
                 p.kill()
                 raise ProcessInterrupted()
     if return_code is None or return_code != 0:
-        raise subprocess.CalledProcessError (
-            return_code if return_code is not None else -1,
-            args,
-            stdout,
-            stderr
+        raise subprocess.CalledProcessError(
+            return_code if return_code is not None else -1, args, stdout, stderr
         )
     return subprocess.CompletedProcess(args, return_code, stdout, stderr)
 
 
 def get_main_am_window():
+    if not QtWidgets:
+        return None
     windows = QtWidgets.QApplication.topLevelWidgets()
     for widget in windows:
         if widget.objectName() == "AddonManager_Main_Window":
             return widget
-    # If there is no main AM window, we may be running unit tests: see if the Test Runner window
-    # exists:
+    # If there is no main AM window, we may be running unit tests: see if the Test
+    # Runner window exists:
     for widget in windows:
         if widget.objectName() == "TestGui__UnitTest":
             return widget
